+++ NEWS.md
@@ -1,3 +1,9 @@
+nodbi 0.9.1.9000
+================
+
+### Changes
+* added warning if DuckDB's JSON extension is not available; see also issue #45
+
 nodbi 0.9.1
 ================